Several bug fixes - Replace size_t with uintmax_t for memory_load() - Define _CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES as 1 - Correct some includes, replace bool with _Bool, correct fprintf, scanf etc formats since the format specs are no longer macros - Move CONSOLE_ADDR_FULL into i8080_cpm_zero_page() since it would have to be static-initialized otherwise. cpm_print_str() now takes the console addr as an argument - MSVC cannot compile const identifiers as array initializers, replaced this with a define which is uglier but it should work

+++ src/emu.h
@@ -18,6 +18,7 @@
 
 #include "i8080/i8080.h"
 #include <stdio.h>
+#include <stdint.h>
 #include "i8080/i8080_predef.h"
 
 I8080_CDECL typedef enum EMU_EXIT_CODE {
@@ -50,7 +51,7 @@ I8080_CDECL_EXTERN const emu_word_t CPM_CONSOLE_ADDR; // = 0x00
 I8080_CDECL_EXTERN const emu_addr_t DEFAULT_START_PA; // = 0x0040
 
 /* Loads a file into memory. Returns number of words read. */
-I8080_CDECL size_t memory_load(const char * file_loc, emu_word_t * memory, emu_addr_t start_loc);
+I8080_CDECL uintmax_t memory_load(const char * file_loc, emu_word_t * memory, emu_addr_t start_loc);
 
 // Initialize an i8080
 I8080_CDECL void emu_init_i8080(i8080 * const cpu);
@@ -77,6 +78,6 @@ I8080_CDECL void emu_set_default_env(i8080 * const cpu);
  * See emu_debug_args to set options on how the output is presented. */
 I8080_CDECL emu_exit_code_t emu_runtime(i8080 * const cpu, _Bool perform_startup_check, emu_debug_args_t * debug_args);
 
-#include "i8080_predef_undef.h"
+#include "i8080/i8080_predef_undef.h"
 
 #endif /* EMU_H */
